move general report functions to report/report.py
